Version 4.10.5
Major changes

+++ PackageFiles/Wix/Make Installer.cmd
@@ -1,7 +1,7 @@
 @echo off
 cls
-set VERSION=4_10_4
-set VERSIONn=4.10.4
+set VERSION=4_10_5c
+set VERSIONn=4.10.5
 
 set WIX_DIR="C:\Program Files (x86)\WiX Toolset v3.7\bin"
 if not exist %WIX_DIR% set WIX_DIR="C:\Program Files\WiX Toolset v3.7\bin"
@@ -14,11 +14,18 @@ set PYRO=%WIX_DIR%\pyro.exe
 
 set WIX_BUILD_DIR=wixBuildDir
 
+set ECLIPSE_PLUGIN_SOURCE=plugins\eclipse
+set CODEWARRIOR_PLUGIN_SOURCE=plugins\codewarrior
+
+rem delete artifact.jar and content.jar
+for %%f in (%ECLIPSE_PLUGIN_SOURCE%\usbdm\artifacts.jar %ECLIPSE_PLUGIN_SOURCE%\usbdm\content.jar) do if exist %%f del /Q %%f
+for %%f in (%CODEWARRIOR_PLUGIN_SOURCE%\usbdm\artifacts.jar %CODEWARRIOR_PLUGIN_SOURCE%\usbdm\content.jar) do if exist %%f del /Q %%f
+
 rem .wxs regenerated Files 
-set REBUILT_WXS=DeviceData.wxs FlashImages.wxs WizardPatchData.wxs Examples.wxs EclipsePlugin_1.wxs EclipsePlugin_2.wxs EclipsePlugin_2.wxs
+set REBUILT_WXS=DeviceData.wxs Stationery.wxs FlashImages.wxs WizardPatchData.wxs Examples.wxs EclipsePlugin_1.wxs EclipsePlugin_2.wxs EclipsePlugin_2.wxs Eclipse_Juno.wxs
 
 set LIGHT_OPTIONS=-ext WixUIExtension -ext WixUtilExtension -sw0204
-set LIGHT_DIRS=-b bin\DeviceData -b bin\FlashImages -b WizardPatches -b Examples -b plugins
+set LIGHT_DIRS=-b bin\DeviceData -b bin\Stationery -b bin\FlashImages -b WizardPatches -b Examples -b %CODEWARRIOR_PLUGIN_SOURCE% -b %ECLIPSE_PLUGIN_SOURCE%
 
 set HEAT_OPTIONS=-srd -ke -gg -sfrag -template fragment -sw5150
 set MSI_FILE=USBDM_%VERSION%_Win
@@ -35,19 +42,24 @@ rem del %MSI_FILE%.msi
 
 if "%1"=="clean" goto finish
 
-if not exist %WIX_BUILD_DIR% mkdir %WIX_BUILD_DIR%
-%HEAT% dir .\bin\DeviceData                 %HEAT_OPTIONS% -cg Cg.DeviceData      -dr D.DeviceData       -out %WIX_BUILD_DIR%\DeviceData.wxs
-%HEAT% dir .\bin\FlashImages                %HEAT_OPTIONS% -cg Cg.FlashImages     -dr D.FlashImages      -out %WIX_BUILD_DIR%\FlashImages.wxs
-%HEAT% dir .\WizardPatches                  %HEAT_OPTIONS% -cg Cg.WizardPatchData -dr D.WizardPatchData  -out %WIX_BUILD_DIR%\WizardPatchData.wxs
-%HEAT% dir .\Examples                       %HEAT_OPTIONS% -cg Cg.Examples        -dr D.Examples         -out %WIX_BUILD_DIR%\Examples.wxs
-%HEAT% dir .\plugins                        %HEAT_OPTIONS% -cg Cg.EclipsePlugin_1 -dr D.EclipsePlugin_1  -out %WIX_BUILD_DIR%\EclipsePlugin_1.wxs
-%HEAT% dir .\plugins                        %HEAT_OPTIONS% -cg Cg.EclipsePlugin_2 -dr D.EclipsePlugin_2  -out %WIX_BUILD_DIR%\EclipsePlugin_2.wxs
-%HEAT% dir .\plugins                        %HEAT_OPTIONS% -cg Cg.EclipsePlugin_3 -dr D.EclipsePlugin_3  -out %WIX_BUILD_DIR%\EclipsePlugin_3.wxs
+if exist %WIX_BUILD_DIR% rmdir /S /Q %WIX_BUILD_DIR%
+mkdir %WIX_BUILD_DIR%
+
+%HEAT% dir .\bin\DeviceData                 %HEAT_OPTIONS% -cg Cg.DeviceData           -dr D.DeviceData                -out %WIX_BUILD_DIR%\DeviceData.wxs
+%HEAT% dir .\bin\Stationery                 %HEAT_OPTIONS% -cg Cg.Stationery           -dr D.Stationery                -out %WIX_BUILD_DIR%\Stationery.wxs
+%HEAT% dir .\bin\FlashImages                %HEAT_OPTIONS% -cg Cg.FlashImages          -dr D.FlashImages               -out %WIX_BUILD_DIR%\FlashImages.wxs
+%HEAT% dir .\WizardPatches                  %HEAT_OPTIONS% -cg Cg.WizardPatchData      -dr D.WizardPatchData           -out %WIX_BUILD_DIR%\WizardPatchData.wxs
+%HEAT% dir .\Examples                       %HEAT_OPTIONS% -cg Cg.Examples             -dr D.Examples                  -out %WIX_BUILD_DIR%\Examples.wxs
+%HEAT% dir .\plugins\codewarrior            %HEAT_OPTIONS% -cg Cg.EclipsePlugin_V10.1  -dr D.CW_V10.1_EclipseDropins   -out %WIX_BUILD_DIR%\EclipsePlugin_1.wxs
+%HEAT% dir .\plugins\codewarrior            %HEAT_OPTIONS% -cg Cg.EclipsePlugin_V10.2  -dr D.CW_V10.2_EclipseDropins   -out %WIX_BUILD_DIR%\EclipsePlugin_2.wxs
+%HEAT% dir .\plugins\codewarrior            %HEAT_OPTIONS% -cg Cg.EclipsePlugin_V10.3  -dr D.CW_V10.3_EclipseDropins   -out %WIX_BUILD_DIR%\EclipsePlugin_3.wxs
+%HEAT% dir .\plugins\codewarrior            %HEAT_OPTIONS% -cg Cg.EclipsePlugin_V10.4  -dr D.CW_V10.4_EclipseDropins   -out %WIX_BUILD_DIR%\EclipsePlugin_4.wxs
+%HEAT% dir .\plugins\eclipse                %HEAT_OPTIONS% -cg Cg.EclipseJunoPlugin    -dr D.EclipseJunoDropins        -out %WIX_BUILD_DIR%\EclipseJunoPluginCDT.wxs
 %CANDLE% -dProductVersion=%VERSIONn% -o %WIX_BUILD_DIR%\ *.wxs %WIX_BUILD_DIR%\*.wxs
 %LIGHT% %LIGHT_OPTIONS% %LIGHT_DIRS% -out %MSI_FILE% %WIX_BUILD_DIR%\*.wixobj
 
-del *.wixpdb
-rmdir /S /Q %WIX_BUILD_DIR%
+rem del *.wixpdb
+rem rmdir /S /Q %WIX_BUILD_DIR%
 
 goto finish
 :doPatch
